SFA-1112 - refactor fixed assets investments note error handling to match recent changes
